Add files via upload
airQual - a measure of the air quality in a certain country; the higher the score, the worse the air quality is, as it means there are more particles in the air. Units: micrograms per cubic meter.

percentEducation - the amount spent on education by the country as a percentage of its GDP

GDP - the gross domestic product of a country, in $USD

covCases - the number of coronavirus cases in a country

covDeaths - the number of coronavirus deaths in a country

avgInc - the average annual household income in a country, in $USD

landlocked - 1 if a country is landlocked, 0 if it isn't landlocked

waterlocked - 1 if a country is entirely surrounded by water (island), 0 if it isn't

popDens - the population density of a country. Units: people per square kilometer.